Extract browser-free @stream-autopilot/core package
Lift the farming engine (scheduler, platform adapters, parsers, watchers,
background controller, tabs primitives) out of the extension into a new
browser-free @stream-autopilot/core package so a future headless CLI can reuse
it without faking chrome/browser globals.

- Split core/tabs.ts: pure *WithBrowser logic, the integrity store, the
  page-context registry and KickWafBlockedError move to core; the thin
  wxt/browser-binding wrappers stay in the extension.
- Add a WatchTabPort and make the adapters' browser-coupled defaults required
  injected deps; the extension wires them in background.ts.
- Thread page-context tab closing through SchedulerTickOptions.stopPageContextTabs
  (browser-free default), removing scheduler's last hidden browser dependency.
- Move DEFAULT_STATE to core/defaults.ts; extension storage.ts reuses it.
- Repoint tests to @stream-autopilot/core and add a boundary guard test that
  fails if core ever imports wxt or a webextension polyfill.

No behavior change. Verified: core typechecks in isolation, 258 tests pass,
Chrome MV3 + Firefox MV2 builds succeed, pnpm -r typecheck clean.
