+++ packages/core/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@stream-autopilot/core",
+  "version": "1.3.0",
+  "type": "module",
+  "private": true,
+  "exports": {
+    "./scheduler": "./src/core/scheduler.ts",
+    "./tablessWatch": "./src/core/tablessWatch.ts",
+    "./twitchIntegrity": "./src/core/twitchIntegrity.ts",
+    "./activityLog": "./src/core/activityLog.ts",
+    "./tabs": "./src/core/tabs.ts",
+    "./defaults": "./src/core/defaults.ts",
+    "./adapter": "./src/platforms/adapter.ts",
+    "./twitch": "./src/platforms/twitch/index.ts",
+    "./twitch/parser": "./src/platforms/twitch/parser.ts",
+    "./twitch/watch": "./src/platforms/twitch/watch.ts",
+    "./kick": "./src/platforms/kick/index.ts",
+    "./kick/parser": "./src/platforms/kick/parser.ts",
+    "./kick/watch": "./src/platforms/kick/watch.ts",
+    "./controller": "./src/background/controller.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@stream-autopilot/shared": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^22.10.2",
+    "typescript": "^5.7.2"
+  }
+}
