+++ packages/core/src/core/activityLog.ts
