+++ packages/core/src/core/scheduler.ts
@@ -10,7 +10,7 @@ import type {
   WatchDecision,
   WatchSession,
 } from "@stream-autopilot/shared/models";
-import { currentManagedPageContextTabs, registerManagedPageContextTabs, stopManagedPageContextTabs } from "./tabs";
+import { currentManagedPageContextTabs, forgetManagedPageContextTabs, registerManagedPageContextTabs, type SchedulerManagedPageContexts } from "./tabs";
 import { appendLog, shouldRecord, type LogLevel } from "@stream-autopilot/shared/logging";
 
 const PLATFORMS: Platform[] = ["twitch", "kick"];
@@ -295,6 +295,13 @@ export interface SchedulerTickResult {
 
 export interface SchedulerTickOptions {
   platforms?: Platform[];
+  // Closes the page-context tabs for the given platforms (browser-bound). Injected
+  // by the runtime so the scheduler stays browser-free; defaults to a no-op that
+  // only updates the retained-context bookkeeping (correct for tabless runtimes).
+  stopPageContextTabs?: (
+    contexts: SchedulerManagedPageContexts,
+    options: { platforms?: Platform[] },
+  ) => Promise<SchedulerManagedPageContexts>;
 }
 
 export async function runSchedulerTick(
@@ -303,6 +310,7 @@ export async function runSchedulerTick(
   adapters: Record<Platform, PlatformAdapter>,
   options: SchedulerTickOptions = {},
 ): Promise<SchedulerTickResult> {
+  const stopPageContextTabs = options.stopPageContextTabs ?? forgetManagedPageContextTabs;
   registerManagedPageContextTabs(state.managedPageContextTabs ?? {});
   let nextState: SchedulerState = {
     ...state,
@@ -345,7 +353,7 @@ export async function runSchedulerTick(
           lastHeartbeatOk: undefined,
         };
         nextState.managedWatchTabs = withoutManagedWatchTab(nextState.managedWatchTabs, platform);
-        nextState.managedPageContextTabs = await stopManagedPageContextTabs(nextState.managedPageContextTabs ?? {}, { platforms: [platform] });
+        nextState.managedPageContextTabs = await stopPageContextTabs(nextState.managedPageContextTabs ?? {}, { platforms: [platform] });
         nextState = addTickEvent(nextState, platform, "info", "Manual watch detected; pausing farming for this platform", enabledLevels);
         continue;
       }
@@ -371,7 +379,7 @@ export async function runSchedulerTick(
           lastHeartbeatOk: undefined,
         };
         nextState.managedWatchTabs = withoutManagedWatchTab(nextState.managedWatchTabs, platform);
-        nextState.managedPageContextTabs = await stopManagedPageContextTabs(nextState.managedPageContextTabs ?? {}, { platforms: [platform] });
+        nextState.managedPageContextTabs = await stopPageContextTabs(nextState.managedPageContextTabs ?? {}, { platforms: [platform] });
         nextState = addTickEvent(nextState, platform, "info", "Automation disabled", enabledLevels);
         continue;
       }
@@ -514,7 +522,7 @@ export async function runSchedulerTick(
             nextState.managedWatchTabs = withoutManagedWatchTab(nextState.managedWatchTabs, platform);
           }
         }
-        nextState.managedPageContextTabs = await stopManagedPageContextTabs(currentManagedPageContextTabs(), { platforms: [platform] });
+        nextState.managedPageContextTabs = await stopPageContextTabs(currentManagedPageContextTabs(), { platforms: [platform] });
         if (settings.autoClaimChannelPoints && adapter.claimChannelPoints) {
           try {
             const claimed = await adapter.claimChannelPoints(decision.channel);
