+++ packages/core/src/platforms/kick/index.ts
@@ -1,8 +1,8 @@
 import type { CategorySelection, ChannelCandidate, ChannelCheck, DropCampaign, DropReward, WatchSession } from "@stream-autopilot/shared/models";
 import type { TablessWatchController } from "../../core/tablessWatch";
 import { logActivity } from "../../core/activityLog";
-import { fetchJsonInPage, fetchKickInBackground, KickWafBlockedError, openPinnedMutedTab, stopWatchTab } from "../../core/tabs";
-import type { PageFetcher, PlatformAdapter, WatchTabOptions } from "../adapter";
+import { KickWafBlockedError } from "../../core/tabs";
+import type { PageFetcher, PlatformAdapter, WatchTabOptions, WatchTabPort } from "../adapter";
 import { kickCandidatesFromCampaign, mergeKickProgress, parseKickCampaigns } from "./parser";
 import { KICK_CLIENT_TOKEN, KickWatcher } from "./watch";
 
@@ -55,12 +55,10 @@ function isKickClaimSuccess(response: KickClaimResponse): boolean {
 // real-Chrome run shows exactly which calls are tabless-capable. The fallback
 // makes this risk-free: farming behaves as before regardless of the result.
 export function createKickFetcher(deps: {
-  background?: (url: string, init?: RequestInit) => Promise<unknown>;
-  pageFetch?: (url: string, init?: RequestInit) => Promise<unknown>;
-} = {}): PageFetcher {
-  const background = deps.background ?? ((url: string, init?: RequestInit) => fetchKickInBackground<unknown>(url, init));
-  const pageFetch = deps.pageFetch ?? ((url: string, init?: RequestInit) =>
-    fetchJsonInPage<unknown>("https://kick.com", url, init, { retainPageContext: { platform: "kick" } }));
+  background: (url: string, init?: RequestInit) => Promise<unknown>;
+  pageFetch: (url: string, init?: RequestInit) => Promise<unknown>;
+}): PageFetcher {
+  const { background, pageFetch } = deps;
   const announced = new Map<string, "background" | "fallback">();
   const report = (host: string, outcome: "background" | "fallback", detail: string): void => {
     const repeat = announced.get(host) === outcome;
@@ -123,9 +121,20 @@ function kickCategoryImage(banner: unknown): string | undefined {
 export class KickAdapter implements PlatformAdapter {
   platform = "kick" as const;
 
+  // Tab-based watch transport; absent when the runtime is tabless-only.
+  private readonly watchTabs?: WatchTabPort;
+
   constructor(
-    private readonly fetcher: PageFetcher = createKickFetcher(),
-  ) {}
+    private readonly fetcher: PageFetcher,
+    deps: { watchTabs?: WatchTabPort } = {},
+  ) {
+    this.watchTabs = deps.watchTabs;
+  }
+
+  private requireWatchTabs(): WatchTabPort {
+    if (!this.watchTabs) throw new Error("KickAdapter: no WatchTabPort configured; tab-based watch is unavailable in this runtime");
+    return this.watchTabs;
+  }
 
   async discoverCampaigns(): Promise<DropCampaign[]> {
     const data = await this.fetcher.fetchJson<unknown>("https://web.kick.com/api/v1/drops/campaigns");
@@ -274,11 +283,11 @@ export class KickAdapter implements PlatformAdapter {
   }
 
   prepareWatchTab(channel: ChannelCandidate, session?: WatchSession, options?: Partial<WatchTabOptions>) {
-    return openPinnedMutedTab(channel, session, options);
+    return this.requireWatchTabs().openPinnedMutedTab(channel, session, options);
   }
 
   stopWatchTab(session: WatchSession, options?: Partial<WatchTabOptions>): Promise<void> {
-    return stopWatchTab(session, options);
+    return this.requireWatchTabs().stopWatchTab(session, options);
   }
 
   // Tabless farming via Kick's viewer WebSocket (see KickWatcher). Reuses this
