+++ packages/extension/package.json
@@ -22,6 +22,7 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
+    "@stream-autopilot/core": "workspace:*",
     "@stream-autopilot/popup-ui": "workspace:*",
     "@stream-autopilot/shared": "workspace:*",
     "@vitejs/plugin-react": "^5.0.0",
