+++ packages/extension/tests/backgroundController.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it, vi } from "vitest";
-import { ALARM_NAME, createBackgroundController } from "../src/background/controller";
+import { ALARM_NAME, createBackgroundController } from "@stream-autopilot/core/controller";
 import type { ChannelCandidate, DropCampaign, DropReward, ExtensionSettings, Platform, SchedulerState } from "@stream-autopilot/shared/models";
 import type { RuntimeSnapshot } from "@stream-autopilot/shared/messages";
 import { DEFAULT_SETTINGS } from "@stream-autopilot/shared/settings";
-import { DEFAULT_STATE } from "../src/core/storage";
-import type { PlatformAdapter } from "../src/platforms/adapter";
-import type { TablessWatchController } from "../src/core/tablessWatch";
+import { DEFAULT_STATE } from "@stream-autopilot/core/defaults";
+import type { PlatformAdapter } from "@stream-autopilot/core/adapter";
+import type { TablessWatchController } from "@stream-autopilot/core/tablessWatch";
 
 const reward = (status: DropReward["status"] = "in_progress"): DropReward => ({
   id: "reward",
